requirements.txt
Django==5.2.5
OCC-Core==7.7.0
Pillow==10.1.0
matplotlib==3.8.2
numpy==1.24.3
ezdxf==1.1.1
